pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "flowrag"
version = "2.0.0"
description = "FlowRAG: Flow-Logic Aware Retrieval-Augmented Generation for Industrial SOP Documents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "FlowRAG Authors"},
]
keywords = ["RAG", "knowledge-graph", "retrieval-augmented-generation", "NLP"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "numpy>=1.21.0",
    "faiss-cpu>=1.7.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
eval = [
    "ragas>=0.1.0",
    "langchain-openai>=0.1.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
]
local = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.4.0",
]

[project.urls]
Repository = "https://github.com/your-org/FlowRAG"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
